requirement.txt
flask
flask-cors
llama-cpp-python
chromadb
openai-whisper
pyaudio
flask_socketio
torchaudio
ctransformers
pygame
soundfile
omegaconf
torchcodec